pkgdown/mathjax.html

twoCoprimary: Sample Size and Power Calculation for Two Co-Primary Endpoints

Comprehensive functions to calculate sample size and power for clinical trials with two co-primary endpoints. The package supports five endpoint combinations: two continuous endpoints (Sozu et al. 2011 doi:10.1080/10543406.2011.551329 ), two binary endpoints using asymptotic methods (Sozu et al. 2010 doi:10.1002/sim.3972 ) and exact methods (Homma and Yoshida 2025 doi:10.1177/09622802251368697 ), mixed continuous and binary endpoints (Sozu et al. 2012 doi:10.1002/bimj.201100221 ), and mixed count and continuous endpoints (Homma and Yoshida 2024 doi:10.1002/pst.2337 ). All methods appropriately account for correlation between endpoints and provide both sample size and power calculation capabilities.

See also

Author

Maintainer: Gosuke Homma my.name.is.gosuke@gmail.com (ORCID)

Authors:

Gosuke Homma my.name.is.gosuke@gmail.com (ORCID)